refactor backend and add more unit tests (#350)
