+++ .vulture_whitelist.py
@@ -1,3 +1,21 @@
 _.sessions  # unused attribute (noxfile.py:7)
 _.reuse_existing_virtualenvs  # unused attribute (noxfile.py:6)
 _.secret_key  # unused attribute (gdbgui/backend.py:104)
+_.reuse_existing_virtualenvs  # unused attribute (noxfile.py:6)
+_.sessions  # unused attribute (noxfile.py:7)
+cover  # unused function (noxfile.py:50)
+lint  # unused function (noxfile.py:78)
+autoformat  # unused function (noxfile.py:94)
+docs  # unused function (noxfile.py:103)
+develop  # unused function (noxfile.py:109)
+serve  # unused function (noxfile.py:118)
+publish  # unused function (noxfile.py:133)
+watch_docs  # unused function (noxfile.py:142)
+build_executable_current_platform  # unused function (noxfile.py:154)
+build_executable_mac  # unused function (noxfile.py:162)
+build_executable_linux  # unused function (noxfile.py:169)
+build_executable_windows  # unused function (noxfile.py:176)
+on_connect  # unused function (tests/test_backend.py:14)
+monkeypatch  # unused variable (tests/test_cli.py:23)
+monkeypatch  # unused variable (tests/test_cli.py:33)
+monkeypatch  # unused variable (tests/test_cli.py:43)
