+++ gdbgui/VERSION.txt
@@ -1,1 +1,1 @@
-0.14.0.0
+0.14.0.1
